Refactor new light system
