+++ src/lib/ecs/systems/LightSystem/visualisations/Bitmap.visual.html
