+++ src/lib/ecs/systems/LightSystem/visualisations/lightSystem.visual.html
